[minor] ensure android and ios work
